src/__init__.py
import os

from dotenv import load_dotenv

load_dotenv()

for env_var in  ['ENV', 'MLFLOW_SERVEUR', 'MLFLOW_REISTRY_NAME']:
        if not(os.getenv(env_var)):
            raise Exception(f"Environment variable {env_var} must be defined.")
        